+++ docs/changes.rst
@@ -11,6 +11,7 @@ Unreleased
 
 **Features**
 
+- New crawler: ``atheistcartoons``
 - New crawler: ``petpeevy``
 - Add status page which for each comic shows when releases are fetched compared
   to the comic's release schedule.
@@ -20,6 +21,7 @@ Unreleased
 - Update crawler ``evilinc`` to work again.
 - Update crawler ``uvod`` to fetch comment too.
 - Update crawler ``gunshow`` to fetch multiple images per release.
+- Update crawler ``questionablecontent`` to work again.
 
 
 1.0.1 (2010-02-23)